Rename grafeas gem
* Rename grafeas-client to grafeas
* Create new grafeas-client gem that depends on grafeas
