+++ grafeas/.repo-metadata.json
@@ -1,7 +1,7 @@
 {
   "name": "grafeas",
   "language": "ruby",
-  "distribution_name": "grafeas-client",
+  "distribution_name": "grafeas",
   "module_name": "Grafeas",
   "module_name_credentials": "Grafeas::V1",
   "env_var_prefix": "GRAFEAS"
