+++ grafeas/.rubocop.yml
@@ -1,0 +1,16 @@
+AllCops:
+  TargetRubyVersion: 2.2
+  Exclude:
+    - "grafeas.gemspec"
+    - "lib/**/*"
+    - "Rakefile"
+    - "test/**/*"
+    - "acceptance/**/*"
+
+Documentation:
+  Enabled: false
+
+Style/StringLiterals:
+  EnforcedStyle: double_quotes
+Style/MethodDefParentheses:
+  EnforcedStyle: require_no_parentheses
