+++ grafeas/.yardopts
@@ -1,5 +1,6 @@
 --no-private
 --title=Container Analysis API
+--exclude _pb\.rb$
 --markup markdown
 --markup-provider redcarpet
