+++ grafeas/Gemfile
@@ -1,0 +1,9 @@
+source "https://rubygems.org"
+
+gemspec
+
+gem "rake", "~> 12.0"
+
+# Pin minitest to 5.11.x to avoid warnings emitted by 5.12.
+# See https://github.com/googleapis/google-cloud-ruby/issues/4110
+gem "minitest", "~> 5.11.3"
