+++ grafeas/acceptance/grafeas/v1/grafeas_service_smoke_test.rb
