+++ grafeas/grafeas.gemspec
@@ -1,15 +1,14 @@
 # -*- ruby -*-
 # encoding: utf-8
-
-require File.expand_path("lib/grafeas/client/version", __dir__)
+require File.expand_path("../lib/grafeas/version", __FILE__)
 
 Gem::Specification.new do |gem|
-  gem.name          = "grafeas-client"
-  gem.version       = Grafeas::Client::VERSION
+  gem.name          = "grafeas"
+  gem.version       = Grafeas::VERSION
 
   gem.authors       = ["Google LLC"]
   gem.email         = "googleapis-packages@google.com"
-  gem.description   = "grafeas-client is the official library for the Grafeas API."
+  gem.description   = "grafeas is the official library for the Grafeas API."
   gem.summary       = "API Client library for the Grafeas API"
   gem.homepage      = "https://github.com/googleapis/googleapis"
   gem.license       = "Apache-2.0"
@@ -20,14 +19,14 @@ Gem::Specification.new do |gem|
                       ["README.md", "AUTHENTICATION.md", "LICENSE", ".yardopts"]
   gem.require_paths = ["lib"]
 
-  gem.required_ruby_version = ">= 2.4"
+  gem.required_ruby_version = ">= 2.0.0"
 
-  gem.add_dependency "grafeas", ">= 0.3" # TODO: bump this when grafeas is released
+  gem.add_dependency "google-gax", "~> 1.7"
+  gem.add_dependency "grpc-google-iam-v1", "~> 0.6.9"
 
-  gem.add_development_dependency "google-style", "~> 1.24.0"
+  gem.add_development_dependency "minitest", "~> 5.10"
   gem.add_development_dependency "redcarpet", "~> 3.0"
+  gem.add_development_dependency "rubocop", "~> 0.64.0"
   gem.add_development_dependency "simplecov", "~> 0.9"
   gem.add_development_dependency "yard", "~> 0.9"
-
-  gem.post_install_message = "grafeas-client is now grafeas, please change the gem name in your dependencies"
 end
