+++ grafeas/lib/grafeas/v1.rb
@@ -33,11 +33,11 @@ module Grafeas
   # 1. [Select or create a Cloud Platform project.](https://console.cloud.google.com/project)
   # 2. [Enable billing for your project.](https://cloud.google.com/billing/docs/how-to/modify-project#enable_billing_for_a_project)
   # 3. [Enable the Container Analysis API.](https://console.cloud.google.com/apis/library/grafeas.googleapis.com)
-  # 4. [Setup Authentication.](https://googleapis.dev/ruby/grafeas-client/latest/file.AUTHENTICATION.html)
+  # 4. [Setup Authentication.](https://googleapis.dev/ruby/grafeas/latest/file.AUTHENTICATION.html)
   #
   # ### Installation
   # ```
-  # $ gem install grafeas-client
+  # $ gem install grafeas
   # ```
   #
   # ### Next Steps
