+++ grafeas/lib/grafeas/v1/doc/google/protobuf/empty.rb
