+++ grafeas/lib/grafeas/v1/doc/grafeas/v1/image.rb
