+++ grafeas/lib/grafeas/v1/grafeas_client.rb
@@ -27,7 +27,7 @@ require "google/gax"
 
 require "grafeas/v1/grafeas_pb"
 require "grafeas/v1/credentials"
-require "grafeas/client/version"
+require "grafeas/version"
 
 module Grafeas
   module V1
@@ -204,7 +204,7 @@ module Grafeas
           updater_proc = credentials.updater_proc
         end
 
-        package_version = Grafeas::Client::VERSION
+        package_version = Grafeas::VERSION
 
         google_api_client = "gl-ruby/#{RUBY_VERSION}"
         google_api_client << " #{lib_name}/#{lib_version}" if lib_name
