+++ grafeas/lib/grafeas/v1/image_pb.rb
