+++ grafeas/lib/grafeas/version.rb
@@ -14,7 +14,5 @@
 
 
 module Grafeas
-  module Client
-    VERSION = "0.3.0".freeze
-  end
+  VERSION = "0.3.0".freeze
 end
